requirements.txt
absl-py==1.4.0
altair==4.2.2
amqp==5.1.1
anyio==3.7.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
astunparse==1.6.3
attrs==23.1.0
awsebcli==3.20.6
bcrypt==4.0.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
blessed==1.20.0
blinker==1.6.2
botocore==1.29.99
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.1
cached-property==1.5.2
cachetools==5.3.0
celery==5.2.7
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
chess==1.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contourpy==1.0.7
cron-descriptor==1.2.35
cryptography==40.0.2
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
Django==2.2
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
dnspython==1.16.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
Eel==0.16.0
entrypoints==0.4
fastapi==0.96.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.11.0
Flask==2.2.3
flatbuffers==23.5.9
fonttools==4.39.3
frozendict==2.3.8
future==0.18.3
gast==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.31
gnews==0.2.7
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
imutils==0.5.4
itsdangerous==2.1.2
jax==0.4.10
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==3.2.0
keras==2.12.0
kiwisolver==1.4.4
kombu==5.2.4
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
multiprocess==0.70.14
multitasking==0.0.11
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.1
pandas==2.0.1
paramiko==3.1.0
pathos==0.3.0
pathspec==0.10.1
Pillow==9.5.0
pipenv==2023.4.20
platformdirs==3.2.0
plotly==5.14.1
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.38
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pydeck==0.8.1b0
PyExecJS==1.5.1
pygame==2.4.0
Pygments==2.15.1
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pymongo==3.12.3
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pypdf==3.8.1
pypng==0.20220715.0
pyrsistent==0.19.3
PySimpleGUI==4.34.0
pytesseract==0.3.10
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
qrcode==7.4.2
QtPy==2.3.1
regex==2023.5.5
reportlab==3.6.12
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
Sastrawi==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.8.5
sgmllib3k==1.0.0
six==1.14.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
starlette==0.27.0
streamlit==1.22.0
streamlit-option-menu==0.3.2
svglib==1.5.1
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
texttable==1.6.7
threadpoolctl==3.1.0
timer==0.1.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==3.4.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
translators==5.7.1
twython==3.9.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
uvicorn==0.22.0
vaderSentiment==3.3.2
validators==0.20.0
vine==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.3
whichcraft==0.6.1
wrapt==1.14.1
xhtml2pdf==0.2.10
yfinance==0.2.17
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
